+++ R/fcc_dictionary.R
@@ -1,14 +1,17 @@
-#' Display FCC variable and descriptions
+#' Dictionary for our datasets
 #'
-#' @param dataset a string matching a dataset
+#' Currently we are providing, Form 477 (f477) data and National Broadband Map
+#' at the broadband services location (NBM_raw) level or
+#' processed at Census Block (NBM_block)
 #'
-#' @return a data frame
-#'
-#' @export
-#'
-#' @examples
-
-fcc_dictionary <- function(dataset) {
-  dict <- cori.data.fcc::dictionary
-  dict[dict[["dataset"]] == dataset, ]
-}
+#' @format ## `fcc_dictionary`
+#' A data frame with 15 rows and 5 columns:
+#' \describe{
+#'   \item{dataset}{name of the dataset}
+#'   \item{var_name}{name of a field/columns}
+#'   \item{var_type}{Data type}
+#'   \item{var_description}{Description either provided by FCC or describing what was our process}
+#'   \item{var_example}{One illustration}
+#' }
+#' @source <https://www.fcc.gov/general/explanation-broadband-deployment-data>
+"fcc_dictionary"