+++ R/get_nbm_available.R
@@ -13,6 +13,7 @@
 #'
 #' @examples
 #' nbm_data <- get_nbm_available()
+#' head(nbm_data)
 
 get_nbm_available <- function(
   get_data_url = paste0("https://broadbandmap.fcc.gov/nbm/map/",
@@ -22,17 +23,35 @@ get_nbm_available <- function(
   get_csv_to_dl <- function(release_file, release_nb) {
     get_data_url <- paste0(get_data_url,
                            release_file[release_nb, "process_uuid"])
-    h <- curl::new_handle()
-    curl::handle_setheaders(h, "User-Agent" = user_agent)
+    # h <- curl::new_handle()
+    # curl::handle_setheaders(h, "User-Agent" = user_agent)
+    #
+    # raw_dat <- curl::curl_fetch_memory(get_data_url)
+    #
+    # csv_to_dl <- jsonlite::fromJSON(rawToChar(raw_dat$content))$data
 
-    raw_dat <- curl::curl_fetch_memory(get_data_url)
+    res <- system(
+      sprintf(
+        paste0("curl '%s' --compressed ",
+               "-H 'User-Agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:131.0) Gecko/20100101 Firefox/131.0' ",
+               "-H 'Accept: text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/png,image/svg+xml,*/*;q=0.8' ",
+               "-H 'Accept-Language: en-US,en;q=0.5' -H 'Accept-Encoding: gzip, deflate, br, zstd' ",
+               "-H 'Connection: keep-alive' -H 'Upgrade-Insecure-Requests: 1' ",
+               "-H 'Sec-Fetch-Dest: document' -H 'Sec-Fetch-Mode: navigate' -H 'Sec-Fetch-Site: none' -H 'Sec-Fetch-User: ?1' ",
+               "-H 'Sec-GPC: 1' -H 'Priority: u=0, i' -H 'Pragma: no-cache' -H 'Cache-Control: no-cache' -H 'TE: trailers'"
+        ),
+        get_data_url
+      ),
+      intern = TRUE)
+    csv_to_dl <- jsonlite::fromJSON(res)[["data"]]
 
-    csv_to_dl <- jsonlite::fromJSON(rawToChar(raw_dat$content))$data
     csv_to_dl[["release"]] <- release_file[release_nb, "filing_subtype"]
     return(csv_to_dl)
   }
 
   release <- cori.data.fcc::get_nbm_release()
+
+  release
 
   big_list <- lapply(seq_len(nrow(release)), get_csv_to_dl, release_file = release)