+++ tests/testthat/test-fcc_dictionary.R
@@ -1,3 +1,0 @@
-test_that("dictionnary is a dataframe", {
-  expect_true(is.data.frame(fcc_dictionary(dataset = "f477")))
-})